fixed width of the customField
