+++ src/components/layout/aside-admin/lastest-requests/LatestRequests.tsx
@@ -1,7 +1,7 @@
 import { memo } from 'react';
 import { BaseText, BoldText } from 'components/reusables/ui/text';
 
-import RequestCard from '../../../reusables/request-card/RequestCard';
+import RequestCard from '../../../reusables/request-card';
 
 const LatestRequests = () => (
   <>
