+++ src/components/pages/requests/lastest-requests/filter-requests/index.tsx
@@ -1,0 +1,129 @@
+import { useState } from 'react';
+import { BoldText } from 'reusables/ui/text';
+import CustomField from 'reusables/custom-field';
+import useCustomField from 'reusables/custom-field/hooks-custom-field/useCustomField';
+
+const testing = [
+  'here',
+  'lakd',
+  'kdgjfksf',
+  'bakfko',
+  'giodsfsl',
+  'sdkfjakdf',
+  'qdooaldkf',
+  'ksdadkfjkdfj akfjsklajfkjkfsfoi',
+];
+
+export default function RequestsFilter() {
+  const [classs, setClasss] = useState('');
+  const [search, setSearch] = useCustomField('');
+  const [user, setUser] = useState<string[]>([]);
+  const [names, setNames] = useState([...testing]);
+  const [value, setValue, list, filterFn] = useCustomField<string[]>(
+    [],
+    testing
+  );
+  const [date, setDate, dateList, dateListFilterFn] = useCustomField<string>(
+    '',
+    testing
+  );
+
+  const addUser = (u: string) => {
+    if (user.includes(u)) {
+      setUser((use) => use.filter((us) => us !== u));
+      return;
+    }
+    setUser([...user, u]);
+  };
+
+  const filterList = (val: string) => {
+    setNames((prev) => prev.filter((v) => v.includes(val)));
+  };
+
+  return (
+    <>
+      <div>
+        <BoldText>Search Keyword:</BoldText>
+        <div className="mt-1">
+          <CustomField onChange={setSearch} field="input" value={search} />
+        </div>
+      </div>
+      <div className="mt-4">
+        <BoldText>Filter By:</BoldText>
+        <div className="mt-1">
+          <CustomField
+            filterFn={filterFn}
+            onSelect={setValue}
+            field="select"
+            value={value}
+          >
+            <CustomField.DropdownWrapper>
+              {list.map((name) => (
+                <CustomField.Dropdown key={name} value={name}>
+                  {name}
+                </CustomField.Dropdown>
+              ))}
+            </CustomField.DropdownWrapper>
+          </CustomField>
+        </div>
+      </div>
+      <div className="mt-4">
+        <BoldText>From Date:</BoldText>
+        <div className="mt-1">
+          <CustomField
+            field="input"
+            filterFn={dateListFilterFn}
+            onChange={setDate}
+            value={date}
+            icon="caretDown"
+          >
+            <CustomField.DropdownWrapper>
+              {dateList.map((name) => (
+                <CustomField.Dropdown key={name} value={name}>
+                  {name}
+                </CustomField.Dropdown>
+              ))}
+            </CustomField.DropdownWrapper>
+          </CustomField>
+        </div>
+      </div>
+      <div className="mt-4">
+        <BoldText>Filter By:</BoldText>
+        <div className="mt-1">
+          <CustomField
+            filterFn={filterList}
+            onSelect={addUser}
+            field="select"
+            value={user}
+          >
+            <CustomField.DropdownWrapper>
+              {names.map((name) => (
+                <CustomField.Dropdown key={name} value={name}>
+                  {name}
+                </CustomField.Dropdown>
+              ))}
+            </CustomField.DropdownWrapper>
+          </CustomField>
+        </div>
+      </div>
+      <div className="mt-4">
+        <BoldText>Select Class:</BoldText>
+        <div className="mt-1">
+          <CustomField
+            onSelect={(v) => setClasss(v)}
+            field="select"
+            value={classs}
+          >
+            <CustomField.DropdownWrapper>
+              {names.map((name) => (
+                <CustomField.Dropdown key={name} value={name}>
+                  {name}
+                </CustomField.Dropdown>
+              ))}
+            </CustomField.DropdownWrapper>
+          </CustomField>
+        </div>
+      </div>
+    </>
+  );
+}
