+++ src/components/reusables/rate-someone/index.tsx
@@ -2,7 +2,7 @@ import { memo, useState } from 'react';
 import { BoldText } from 'components/reusables/ui/text';
 import CustomField from 'components/reusables/custom-field';
 import useCustomField from 'components/reusables/custom-field/hooks-custom-field/useCustomField';
-import StarRatings from 'components/reusables/star-ratings/StarRatings';
+import StarRatings from 'components/reusables/star-ratings';
 
 const RateSomeone = () => {
   const [user, setUser, list, filterFn] = useCustomField(
